--- packages/core/src/components/BreadCrumbs/BreadCrumbs.stories.mdx
@@ -1,17 +1,0 @@
-import { BreadCrumbs } from './BreadCrumbs.tsx';
-import { Preview, Story, Meta, Props } from '@storybook/addon-docs/blocks';
-import * as stories from './BreadCrumbs.stories';
-
-<Meta title="Core" component={BreadCrumbs} />
-
-# BreadCrumbs
-
-### Breadcrumbs allow users to make selections from a range of values.
-
-<Preview withToolbar>
-    <Story name="BreadCrumbs">{stories.Basic()}</Story>
-</Preview>
-
-### BreadCrumbsTheme
-
-<Props of={stories.ThemeInterface} />
