--- packages/core/src/components/BreadCrumbs/index.ts
@@ -1,3 +1,0 @@
-import { BreadCrumbs } from './BreadCrumbs';
-
-export default BreadCrumbs;
