+++ packages/core/src/components/Breadcrumb/index.ts
@@ -1,0 +1,1 @@
+export { Breadcrumb as default } from './Breadcrumb';
