--- packages/theme/src/core/breadCrumbs/index.ts
@@ -1,9 +1,0 @@
-import colors from '../colors';
-import { BreadCrumbsTheme } from './types';
-
-const breadCrumbs: BreadCrumbsTheme = {
-    textColor: colors.white,
-    bgColor: colors.professionalNavy
-};
-
-export default breadCrumbs;
