+++ packages/theme/src/core/breadcrumb/index.ts
@@ -1,0 +1,9 @@
+import colors from '../colors';
+import { BreadcrumbTheme } from './types';
+
+const breadcrumb: BreadcrumbTheme = {
+    textColor: colors.white,
+    bgColor: colors.professionalNavy
+};
+
+export default breadcrumb;
